+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Remove `CircleCI::CoverageReporter::Client#configuration`
+
+### Fixed
+- Run each example in sandbox
 
 ## [0.2.0] - 2017-03-14
 ### Added
@@ -21,17 +26,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.2] - 2017-03-12
 ### Fixed
-
 - Set `nil` to previous_build_number if CIRCLE_PREVIOUS_BUILD_NUM environment variable is empty
 
 ## [0.1.1] - 2017-03-12
 ### Fixed
-
 - Ignore base_build if it is on master branch
 
 ## 0.1.0 - 2017-03-12
 ### Added
-
 - Initial release
 
 [Unreleased]: https://github.com/increments/circleci-coverage_reporter/compare/v0.2.0...HEAD